+++ packages/terra-form-select/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added `isFilterStyle` prop to native-select-field.
+
 ## 6.40.0 - (April 27, 2023)
 
 * Added